Split up dwds.dart file
